requirements.txt
langchain
langchain-google-genai
langchain-community
yfinance
pandas
numpy
streamlit
python-dotenv
requests
matplotlib
plotly
scipy